requirements.txt
azure-identity>=1.15.0
requests>=2.31.0
